Tratando erros

+++ api/routes/alunoRoute.js
@@ -1,4 +1,5 @@
 const alunoRoute = require('express').Router();
+const e = require('express');
 const aluno = require('../models/alunoModel');
 
 //GET ALL - Lista todos os alunos cadastrados
@@ -34,12 +35,21 @@ alunoRoute.get('/aluno/id=:id', async (req, res) => {
 //POST - Cadastra um novo aluno, passando os dados pelo body
 alunoRoute.post('/add-aluno', async (req, res) => {
     try {
-        if(Object.values(req.body).length===0){
-            res.json({mensagem: 'Erro!'}); 
+
+        const nome = req.body.nome
+        const idade = req.body.idade
+        const email = req.body.email
+        const telefone = req.body.telefone
+
+        if(nome === undefined || idade === undefined || email === undefined|| telefone===undefined ){
+            res.json({mensagem: 'Erro! Alguns campos não foram definidos!'});  
+        }else{
+        if(await aluno.findOne({"email": email})!=null){
+            res.json({mensagem: 'Erro no cadastro! Email já cadastrado!'}); 
         }else{
         await aluno.create(req.body);
         res.json({mensagem: 'Aluno cadastrado com sucesso!'});
-        }
+        }}
     } catch (error) {
         res.json({mensagem: 'Erro no cadastro!'});
     }
@@ -48,36 +58,59 @@ alunoRoute.post('/add-aluno', async (req, res) => {
 //PUT - Atualiza um aluno específico pelo email, passando os dados pelo body
 alunoRoute.put('/update-aluno/email=:email', async (req, res) => {
     try {
-        if(Object.values(req.body).length===0){
-            res.json({mensagem: 'Erro!'}); 
+
+        const nome = req.body.nome
+        const idade = req.body.idade
+        const email = req.body.email
+        const telefone = req.body.telefone
+
+        if(nome === undefined && idade === undefined && email === undefined && telefone===undefined ){
+            res.json({mensagem: 'Erro ao atualizar! Nenhum parâmetro encontrado!'}); 
+        }else{
+        if(await aluno.findOne({"email": email})!=null){
+                res.json({mensagem: 'Erro ao atualizar! Email já existe!'});
         }else{
         await aluno.findOneAndUpdate({"email": req.params.email}, req.body, {new:true});
         res.json({mensagem: 'Aluno atualizado com sucesso!'});
-        }
+        }}
     } catch (error) {
-        res.json({mensagem: 'Erro na atualiazação!'});
+        res.json({mensagem: 'Erro na atualização!'});
     }
 });
 
 //PUT - Atualiza um aluno específico pelo id, passando os dados pelo body
 alunoRoute.put('/update-aluno/id=:id', async (req, res) => {
     try {
-        if(Object.values(req.body).length===0){
-            res.json({mensagem: 'Erro!'}); 
+        
+        const nome = req.body.nome
+        const idade = req.body.idade
+        const email = req.body.email
+        const telefone = req.body.telefone
+
+        if(nome === undefined && idade === undefined && email === undefined && telefone===undefined ){
+            res.json({mensagem: 'Erro ao atualizar! Nenhum parâmetro encontrado!'}); 
+        }else{
+        if(await aluno.findOne({"email": email})!=null){
+            res.json({mensagem: 'Erro ao atualizar! Email já existe!'});
         }else{
         await aluno.findOneAndUpdate({"_id": req.params.id}, req.body, {new:true});
         res.json({mensagem: 'Aluno atualizado com sucesso!'});
-        }
+        }}
     } catch (error) {
-        res.json({mensagem: 'Erro na atualiazação!'});
+        res.json({mensagem: 'Erro na atualização!'});
     }
 });
 
 //DELETE - Remove um aluno específico pelo email
 alunoRoute.delete('/delete-aluno/email=:email', async (req, res) => {
     try {
+      if(await aluno.findOne({"email": req.params.email})!=null){
         await aluno.deleteOne({"email": req.params.email});
-        res.json({mensagem: 'Aluno removido com sucesso!'}); 
+        res.json({mensagem: 'Aluno removido com sucesso!'});
+      }else{
+        res.json({mensagem: 'Erro ao remover! Aluno não existe!'});
+      } 
+    
     } catch (error) {
         res.json({mensagem: 'Erro na exclusão!'});
     }
@@ -86,8 +119,12 @@ alunoRoute.delete('/delete-aluno/email=:email', async (req, res) => {
 //DELETE - Remove um aluno específico pelo id
 alunoRoute.delete('/delete-aluno/id=:id', async (req, res) => {
     try {
-        await aluno.deleteOne({"_id": req.params.id});
-        res.json({mensagem: 'Aluno removido com sucesso!'}); 
+        if(await aluno.findOne({"_id": req.params.id})!=null){
+            await aluno.deleteOne({"_id": req.params.id});
+            res.json({mensagem: 'Aluno removido com sucesso!'});
+          }else{
+            res.json({mensagem: 'Erro ao remover! Aluno não existe!'});
+          } 
     } catch (error) {
         res.json({mensagem: 'Erro na exclusão!'});
     }
